bugs fixed

+++ ECommerce.Application/DTOs/Customer/CustomerReviewDto.cs
@@ -3,6 +3,6 @@
 public sealed record CustomerReviewDto
 {
     public int ProductId { get; init; }
-    public byte Rating { get; init; }
+    public int Rating { get; init; }
     public string Comment { get; init; }
 }
