+++ ECommerce.Application/DTOs/Product/ProductReviewDto.cs
@@ -4,6 +4,6 @@ public sealed class ProductReviewDto
 {
     public string CustomerFullName { get; init; }
     public string ProductName { get; init; }
-    public byte Rating { get; init; }
+    public int Rating { get; init; }
     public string Comment { get; init; }
 }
