+++ ECommerce.Infrastructure/Persistence/Configurations/PaymentStatusConfiguration.cs
@@ -26,9 +26,9 @@ public class PaymentStatusConfiguration : IEntityTypeConfiguration<PaymentStatus
             .IsUnique()
             .HasDatabaseName("uq_aram_payment_status_name");
 
-        builder.HasData(new PaymentStatus { Id = (int)Domain.Enums.PaymentStatus.Pending, Name = Domain.Enums.PaymentStatus.Pending.GetDescription() },
-                        new PaymentStatus { Id = (int)Domain.Enums.PaymentStatus.Paid, Name = Domain.Enums.PaymentStatus.Paid.GetDescription() },
-                        new PaymentStatus { Id = (int)Domain.Enums.PaymentStatus.Refunded, Name = Domain.Enums.PaymentStatus.Refunded.GetDescription() },
-                        new PaymentStatus { Id = (int)Domain.Enums.PaymentStatus.Cancelled, Name = Domain.Enums.PaymentStatus.Cancelled.GetDescription() });
+        builder.HasData(new PaymentStatus { Id = Domain.Enums.PaymentStatus.Pending, Name = Domain.Enums.PaymentStatus.Pending.GetDescription() },
+                        new PaymentStatus { Id = Domain.Enums.PaymentStatus.Paid, Name = Domain.Enums.PaymentStatus.Paid.GetDescription() },
+                        new PaymentStatus { Id = Domain.Enums.PaymentStatus.Refunded, Name = Domain.Enums.PaymentStatus.Refunded.GetDescription() },
+                        new PaymentStatus { Id = Domain.Enums.PaymentStatus.Cancelled, Name = Domain.Enums.PaymentStatus.Cancelled.GetDescription() });
     }
 }
